+++ app/package.json
@@ -10,7 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@moondreamsdev/dreamer-ui": "^1.7.10-test.47",
+    "@moondreamsdev/dreamer-ui": "^1.7.10-test.48",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-router-dom": "^7.8.2"
