+++ lib/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@moondreamsdev/dreamer-ui",
-  "version": "1.7.10-test.47",
+  "version": "1.7.10-test.49",
   "publishConfig": {
     "access": "public"
   },
